// src/app/models/coordenada.ts
export interface Coordenada {
  id?: number;
  gradosLatitud?: string;
  minutosLatitud?: string;
  segundosLatitud?: string;
  gradosLongitud?: string;
  minutosLongitud?: string;
  segundosLongitud?: string;
  altitud?: string;
  nmeaLatitud?: string;
  nmeaLongitud?: string;
  utmX?: string;
  utmY?: string;
  utmHuso?: string;
  utmBanda?: string;
  xy_x?: string;
  xy_y?: string;
  xy_z?: string;
}